Add Fedora4 role
This adds support for a fedora4 role which brings up a default Fedora 4
instance via Tomcat 7.  The version of Fedora to be used is determined
by setting either the "project_fedora_version" or "fedora_version"
variables to the Fedora 4 version required.  Thus far, only versions
4.5.1 and 4.7.1 are supported.  If no version is specified explicitly,
it defaults to 4.7.1.

The installed Fedora 4 runs at the /fcrepo endpoint of the Tomcat 7
server.