Add Fedora4 role (#20)
* Add Fedora4 role

This adds support for a fedora4 role which brings up a default Fedora 4
instance via Tomcat 7.  The version of Fedora to be used is determined
by setting either the "project_fedora_version" or "fedora_version"
variables to the Fedora 4 version required.  Thus far, only versions
4.5.1 and 4.7.1 are supported.  If no version is specified explicitly,
it defaults to 4.7.1.

The installed Fedora 4 runs at the /fcrepo endpoint of the Tomcat 7
server.

* Improve modeshape configuration logic

After clarification from Yinlin Chen about default locations for the
modeshape configuration file, the logic for locating it becomes
simpler. This also means we cover all existing versions of Fedora 4,
not just the two defined in the previous version of the code.

* Support flexible modeshape configuration

This change allows more flexible configuration of the modeshape
configuration.  It permits a local modeshape "repository.json" file to
be supplied, which will be used in preference to any included modeshape
configuration setting specified via "project_modeshape_config".

Thus, if a local file specified by "fedora_local_modeshape_config_file"
exists, it is copied (as "fedora_remote_modeshape_config_file") to the
system being provisioned and Fedora 4 is directed to use this file.
Otherwise, the default configuration shipped with the Fedora 4 WAR file
directory specified via "project_modeshape_config" is used instead.
This latter configuration defaults to "file-simple" if
"project_modeshape_config" is not explicitly set.  Note, if using this
configuration approach, ensure that the configuration directory
specified by "project_modeshape_config" actually exists for the
"fedora_version" being deployed.